move network selector to folder
